requirements.txt
mutagen==1.31
Pillow==3.2.0
